requirements.txt
annotated-types==0.6.0
anyio==4.6.2
click==8.1.7
fastapi==0.112.2
h11==0.14.0
idna==3.7
pip==24.2
pydantic==2.8.2
pydantic_core==2.20.1
setuptools==75.1.0
sniffio==1.3.0
starlette==0.38.2
typing_extensions==4.11.0
uvicorn==0.30.6
wheel==0.44.0
pycryptodome==3.10.1
requests==2.32.3
pandas
tqdm==4.66.5
imblearn
beautifulsoup4==4.12.3
bs4==0.0.2
Pyrebase4
